// rs-clone/client/src/enums/app-text-content.ts
enum AppTextContent {
  LOGIN = 'Логин',
  NICK_NAME = 'Никнейм (публичный)',
  PASSWORD = 'Пароль',
  AVATAR = 'Аватар (в формате png, jpg, jpeg, до 500Кб)',
  AVATAR_FORMAT_ERROR = 'Недопустимый формат',
  AVATAR_SIZE_ERROR = 'Недопустимый размер',
  CHOOSE_AVATAR = 'Выберите файл',
  SUBMIT = 'Готово'
}
export default AppTextContent;